migraciones

+++ ArtGallery/Data/Models/Museum.cs
@@ -5,8 +5,8 @@ namespace ArtGallery.Models {
 	public class Museum {
 		public string? City { get; set; }
 		public string? State { get; set; }
-		public int? Latitude { get; set; }
-		public int? Longitude { get; set; }
+		public float? Latitude { get; set; }
+		public float? Longitude { get; set; }
 		public int? Inauguration { get; set; }
 		public required string Slug { get; set; }
 		public required string Name { get; set; }
